chore: abstract college form page
